Rename module to 'ethpm'
