+++ tests/ethpm/invalidLockfile.json
