+++ tests/ethpm/validLockfile.json
