+++ tox.ini
@@ -20,4 +20,4 @@ basepython =
 [testenv:flake8]
 basepython=python
 deps=flake8
-commands=flake8 {toxinidir}/tests/erc190 {toxinidir}/erc190 
+commands=flake8 {toxinidir}/tests/ethpm {toxinidir}/ethpm
